requirements.txt
urllib3
requests
papermill
ijson
ipykernel 
setuptools
nbopen
gitpython
giturlparse
colorama
